expand the error reporting when sync'ing

+++ modules/gitbox/files/cgi-bin/sync-repo.cgi
@@ -245,14 +245,17 @@ elif 'repository' in data and 'name' in data['repository']:
             out = subprocess.check_output(["git", "fetch", "--prune"])
             log += "[%s] [%s.git]: Git fetch succeeded\n" % (time.strftime("%c"), reponame)
             try:
-                os.unlink("/x1/gitbox/broken/%s.txt" % cfg.repo_name)
+                if os.path.exists("/x1/gitbox/broken/%s.txt" % cfg.repo_name):
+                    os.unlink("/x1/gitbox/broken/%s.txt" % cfg.repo_name)
             except:
                 pass
         except subprocess.CalledProcessError as err:
             broken = True
             log += "[%s] [%s.git]: Git fetch failed: %s\n" % (time.strftime("%c"), reponame, err.output)
             with open("/x1/gitbox/broken/%s.txt" % reponame, "w") as f:
-                f.write("BROKEN AT %s\n" % time.strftime("%c"))
+                f.write("BROKEN AT %s\n\nOutput:\n" % time.strftime("%c"))
+                f.write("Return code: %s\nText output:\n" % err.returncode)
+                f.write(err.output)
                 f.close()
             
             # Send an email to users@infra.a.o with the bork
